Fix common header multiple implementation for Arduino
Fix ESP-IDF compilation error in debugger.cpp

+++ src/Debug/debugger.cpp
@@ -357,7 +357,7 @@ void Debugger::dumpLocals(Module *m) const {
     dprintf(this->socket, R"({"count":%u,"locals":[)", 0);
     //    fflush(stdout);  // FIXME: this is needed for ESP to propery print
     char _value_str[256];
-    for (size_t i = 0; i < f->block->local_count; i++) {
+    for (uint32_t i = 0; i < f->block->local_count; i++) {
         auto v = &m->stack[m->fp + i];
         switch (v->value_type) {
             case I32:
@@ -382,7 +382,7 @@ void Debugger::dumpLocals(Module *m) const {
                          v->value_type, v->value.uint64);
         }
 
-        dprintf(this->socket, "{%s, \"index\":%lu}%s", _value_str,
+        dprintf(this->socket, "{%s, \"index\":%u}%s", _value_str,
                 i + f->block->type->param_count,
                 (i + 1 < f->block->local_count) ? "," : "");
     }
